Updated `font_character_validate`

+++ game/source/text/font_fallback.cpp
@@ -16,7 +16,7 @@ bool fallback_font_get_character(e_utf32 utf_character, s_font_character const**
 	ASSERT(g_fallback_font_header);
 	ASSERT(g_fallback_font_header->location_table_offset + g_fallback_font_header->location_table_count * sizeof(dword) <= sizeof(k_fallback_font_data));
 
-	if (VALID_INDEX(utf_character, g_fallback_font_header->location_table_count))
+	if (VALID_INDEX((long)utf_character, g_fallback_font_header->location_table_count))
 	{
 		dword location = *(dword*)offset_pointer(k_fallback_font_data, g_fallback_font_header->location_table_offset + sizeof(dword) * utf_character);
 		if (location != NONE)
